+++ Describer/SwaggerPhpDescriber.php
@@ -14,6 +14,7 @@ namespace Nelmio\ApiDocBundle\Describer;
 use Doctrine\Common\Annotations\Reader;
 use EXSyst\Component\Swagger\Swagger;
 use Nelmio\ApiDocBundle\Annotation\Operation;
+use Nelmio\ApiDocBundle\Annotation\Security;
 use Nelmio\ApiDocBundle\SwaggerPhp\AddDefaults;
 use Nelmio\ApiDocBundle\SwaggerPhp\ModelRegister;
 use Nelmio\ApiDocBundle\Util\ControllerReflector;
@@ -45,7 +46,7 @@ final class SwaggerPhpDescriber extends ExternalDocDescriber implements ModelReg
             $analysis->process($this->getProcessors());
             $analysis->validate();
 
-            return json_decode(json_encode($analysis->swagger));
+            return json_decode(json_encode($analysis->swagger), true);
         }, $overwrite);
     }
 
@@ -93,6 +94,7 @@ final class SwaggerPhpDescriber extends ExternalDocDescriber implements ModelReg
             $nestedContext->nested = true;
             $implicitAnnotations = [];
             $tags = [];
+            $security = [];
             foreach ($annotations as $annotation) {
                 $annotation->_context = $context;
                 $this->updateNestedAnnotations($annotation, $nestedContext);
@@ -121,6 +123,13 @@ final class SwaggerPhpDescriber extends ExternalDocDescriber implements ModelReg
                     continue;
                 }
 
+                if ($annotation instanceof Security) {
+                    $annotation->validate();
+                    $security[] = [$annotation->name => []];
+
+                    continue;
+                }
+
                 if ($annotation instanceof SWG\Tag) {
                     $annotation->validate();
                     $tags[] = $annotation->name;
@@ -141,7 +150,7 @@ final class SwaggerPhpDescriber extends ExternalDocDescriber implements ModelReg
 
             foreach ($httpMethods as $httpMethod) {
                 $annotationClass = $operationAnnotations[$httpMethod];
-                $operation = new $annotationClass(['_context' => $context, 'path' => $path, 'value' => $implicitAnnotations, 'tags' => $tags]);
+                $operation = new $annotationClass(['_context' => $context, 'path' => $path, 'value' => $implicitAnnotations, 'tags' => $tags, 'security' => $security]);
                 $analysis->addAnnotation($operation, null);
             }
         }