+++ Tests/Functional/SwaggerUiTest.php
@@ -47,6 +47,7 @@ class SwaggerUiTest extends WebTestCase
             '/api/dummies/{id}' => $expected['paths']['/api/dummies/{id}'],
             '/test/test/' => ['get' => [
                 'responses' => ['200' => ['description' => 'Test']],
+                'security' => [],
             ]],
         ];
         $expected['definitions'] = ['Dummy' => $expected['definitions']['Dummy']];